# backend/application/src/main/resources/api/v1/specs/openapi.yaml
openapi: 3.0.3
info:
  title: Strategy & Planning API
  version: 0.1.0
  description: |
    Strategy & Planning toolkit API. Provides endpoints for the Case Study Builder,
    Category Analysis Builder, and RFP Outline Generator tools. Authentication via Clerk SSO (Bearer JWT).
servers:
  - url: /api/v1
    description: V1 base path (DOCUMENTATION ONLY — the openapi-generator-maven-plugin
      does NOT prepend this to generated Spring routes. Path keys below carry the
      `/api/v1/` prefix verbatim. See canonical-openapi-rules.md → "/api/v1 prefix").
security:
  - bearerAuth: [ ]
...
paths:
...
  /api/v1/case-studies:
    get:
      tags: [ CaseStudies ]
      operationId: listCaseStudies
      summary: List all case studies for the authenticated user
      description: Returns all case studies created by the authenticated user, ordered by creation date descending.
      security: [ { bearerAuth: [ ] } ]
      responses:
        "200":
          description: List of case study summaries for the authenticated user.
          content:
            application/json:
              schema:
                type: array
                description: Ordered list of case study summaries.
                items:
                  $ref: "#/components/schemas/CaseStudySummaryV1"
        "401":
          description: The authorization token not provided or expired.
        "403":
          description: The authorization token not correct or user not authorized.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AppApiExceptionResponseV1"
        "500":
          description: Exception occurred while listing case studies.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AppApiExceptionResponseV1"

    post:
      tags: [ CaseStudies ]
      operationId: createCaseStudy
      summary: Create a new case study and generate its Google Slides deck
      description: |
        Creates a new case study with the provided form fields and immediately generates a Google
        Slides deck by copying the tokenized master template into the user's Drive and filling all
        template tokens. Requires the user to be signed in with Google (via Clerk) with Drive and
        Slides access. On success the record is persisted with status GENERATED and the deck URL
        is returned in slidesUrl.
      security: [ { bearerAuth: [ ] } ]
      requestBody:
        required: true
        description: Case study form fields. Title is required; all other fields are optional pre-fill.
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCaseStudyRequestV1"
      responses:
        "201":
          description: Case study created and its Google Slides deck generated successfully.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CaseStudyV1"
        "400":
          description: Request validation failed.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AppValidationExceptionResponseV1"
        "401":
          description: The authorization token not provided or expired.
        "403":
          description: The authorization token not correct or user not authorized.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AppApiExceptionResponseV1"
        "500":
          description: Exception occurred while creating the case study.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AppApiExceptionResponseV1"
        "502":
          description: >-
            Google deck generation failed — the user is not connected to Google, has not granted
            Drive & Slides access, or a Google API call failed.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AppApiExceptionResponseV1"
  /api/v1/case-studies/drafts:
    post:
      tags: [ CaseStudies ]
      operationId: draftCaseStudy
      summary: AI-draft case study fields from uploaded source documents
      description: |
        Accepts up to 10 source documents (PDF, PowerPoint, Excel, or CSV), extracts their
        readable content — including tabular data from Excel workbooks — and uses the AI
        drafting engine to pre-fill every case study form field. The result is meant to be
        reviewed and edited by the user before submission; metrics and quotes are AI-extracted
        and must be verified against the source documents.
      security: [ { bearerAuth: [ ] } ]
      requestBody:
        required: true
        description: Source documents to draft from. Between 1 and 10 files.
        content:
          multipart/form-data:
            schema:
              type: object
              description: Multipart payload carrying the uploaded source documents.
              required: [ files ]
              properties:
                files:
                  type: array
                  description: Uploaded source documents (PDF, PPTX, PPT, XLSX, XLS, or CSV). Maximum 10 files.
                  minItems: 1
                  maxItems: 10
                  items:
                    type: string
                    format: binary
                    description: One uploaded source document.
      responses:
        "200":
          description: Drafted case study fields for review.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CaseStudyDraftV1"
        "400":
          description: Request validation failed (no files, more than 10 files, or an unsupported/unreadable file).
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AppApiExceptionResponseV1"
        "401":
          description: The authorization token not provided or expired.
        "403":
          description: The authorization token not correct or user not authorized.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AppApiExceptionResponseV1"
        "502":
          description: AI drafting engine not configured or the drafting call failed.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AppApiExceptionResponseV1"
        "500":
          description: Exception occurred while drafting the case study.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AppApiExceptionResponseV1"
  /api/v1/case-studies/{id}:
    get:
      tags: [ CaseStudies ]
      operationId: getCaseStudy
      summary: Get a case study by ID
      description: Returns the full detail of a case study belonging to the authenticated user.
      security: [ { bearerAuth: [ ] } ]
      parameters:
        - name: id
          in: path
          required: true
          description: The unique identifier of the case study.
          schema:
            type: integer
            format: int64
      responses:
        "200":
          description: Case study detail.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CaseStudyV1"
        "401":
          description: The authorization token not provided or expired.
        "403":
          description: The authorization token not correct or user not authorized.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AppApiExceptionResponseV1"
        "404":
          description: Case study not found or does not belong to the authenticated user.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AppApiExceptionResponseV1"
        "500":
          description: Exception occurred while retrieving the case study.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AppApiExceptionResponseV1"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    GenerationStatusV1:
      type: string
      description: |
        Lifecycle status of a generated document.
        DRAFT — saved but not yet submitted for generation.
        SUBMITTED — submitted for Google Slides generation.
        GENERATED — Google Slides deck created in the user's Drive.
      enum: [ DRAFT, SUBMITTED, GENERATED ]
...
    CaseStudySummaryV1:
      type: object
      additionalProperties: false
      description: Lightweight summary of a case study for list views.
      required: [ id, title, status, createdAt ]
      properties:
        id:
          type: integer
          format: int64
          description: Unique identifier of the case study.
        title:
          type: string
          description: Title of the case study.
        clientName:
          type: string
          nullable: true
          description: Name of the client featured in the case study.
        industry:
          type: string
          nullable: true
          description: Industry or sector of the client.
        status:
          description: Lifecycle state of the case study's deck generation.
          allOf:
            - $ref: "#/components/schemas/GenerationStatusV1"
        createdAt:
          type: string
          format: date-time
          description: UTC timestamp when the case study was created.
    CaseStudyV1:
      type: object
      additionalProperties: false
      description: Full detail of a case study including all form fields and generation metadata.
      required: [ id, title, status, createdAt ]
      properties:
        id:
          type: integer
          format: int64
          description: Unique identifier of the case study.
        title:
          type: string
          description: Title of the case study.
        clientName:
          type: string
          nullable: true
          description: Name of the client featured in the case study.
        industry:
          type: string
          nullable: true
          description: Industry or sector of the client.
        challenge:
          type: string
          nullable: true
          description: Description of the client's challenge or problem statement.
        solution:
          type: string
          nullable: true
          description: Description of the solution or approach delivered.
        results:
          type: string
          nullable: true
          description: Outcomes and results achieved for the client.
        keyMetrics:
          type: string
          nullable: true
          description: Key quantitative metrics demonstrating impact (e.g. "40% cost reduction, 2x pipeline growth").
        timeline:
          type: string
          nullable: true
          description: Project timeline or duration (e.g. "6 months", "Q1-Q3 2025").
        testimonial:
          type: string
          nullable: true
          description: Client quote or testimonial to feature in the deck.
        sourceFileName:
          type: string
          nullable: true
          description: Comma-separated names of the uploaded source files used for auto-drafting, if provided.
        templateUrl:
          type: string
          nullable: true
          description: URL of the Google Slides tokenized template used for generation.
        slidesUrl:
          type: string
          nullable: true
          description: URL of the generated Google Slides deck in the user's Drive, when created.
        status:
          description: Lifecycle state of the case study's deck generation.
          allOf:
            - $ref: "#/components/schemas/GenerationStatusV1"
        createdAt:
          type: string
          format: date-time
          description: UTC timestamp when the case study was created.
    CreateCaseStudyRequestV1:
      type: object
      additionalProperties: false
      description: Request payload for creating a new case study. Title is required; all other fields are optional pre-fill.
      required: [ title ]
      properties:
        title:
          type: string
          minLength: 1
          maxLength: 255
          description: Title of the case study.
        clientName:
          type: string
          maxLength: 255
          nullable: true
          description: Name of the client featured in the case study.
        industry:
          type: string
          maxLength: 255
          nullable: true
          description: Industry or sector of the client.
        challenge:
          type: string
          maxLength: 5000
          nullable: true
          description: Description of the client's challenge or problem statement.
        solution:
          type: string
          maxLength: 5000
          nullable: true
          description: Description of the solution or approach delivered.
        results:
          type: string
          maxLength: 5000
          nullable: true
          description: Outcomes and results achieved for the client.
        keyMetrics:
          type: string
          maxLength: 2000
          nullable: true
          description: Key quantitative metrics demonstrating impact.
        timeline:
          type: string
          maxLength: 255
          nullable: true
          description: Project timeline or duration.
        testimonial:
          type: string
          maxLength: 2000
          nullable: true
          description: Client quote or testimonial.
        sourceFileName:
          type: string
          maxLength: 2000
          nullable: true
          description: Comma-separated names of the uploaded source files used for auto-drafting, if provided.
    CaseStudyDraftV1:
      type: object
      additionalProperties: false
      description: |
        AI-drafted case study fields extracted from the uploaded source documents.
        Every field is nullable — the model only fills what the documents support.
      required: [ sourceFileNames ]
      properties:
        title:
          type: string
          nullable: true
          description: Drafted title of the case study.
        clientName:
          type: string
          nullable: true
          description: Drafted client name.
        industry:
          type: string
          nullable: true
          description: Drafted industry or sector.
        challenge:
          type: string
          nullable: true
          description: Drafted challenge or problem statement.
        solution:
          type: string
          nullable: true
          description: Drafted solution or approach delivered.
        results:
          type: string
          nullable: true
          description: Drafted outcomes and results achieved.
        keyMetrics:
          type: string
          nullable: true
          description: Drafted key quantitative metrics, extracted from document data where available.
        timeline:
          type: string
          nullable: true
          description: Drafted project timeline or duration.
        testimonial:
          type: string
          nullable: true
          description: Drafted client quote or testimonial found in the documents.
        sourceFileNames:
          type: string
          description: Comma-separated names of the uploaded source files the draft was produced from.
...
    AppValidationExceptionResponseV1:
      type: object
      additionalProperties: false
      description: Representation of validation exception occurred in REST API format.
      required: [ errors, timestamp, correlationId ]
      properties:
        errors:
          description: The list of validation errors occurred.
          type: array
          items:
            $ref: "#/components/schemas/FieldToErrorResponseV1"
        timestamp:
          type: string
          format: date-time
          description: The timestamp exception occurred.
        correlationId:
          type: string
          description: The log stream UUID.
    FieldToErrorResponseV1:
      type: object
      additionalProperties: false
      description: Representation of validation error bound to a field or semantic request part.
      required: [ code, field, error ]
      properties:
        code: { type: string, description: "The code of validation error occurred." }
        field: { type: string, description: "The field or semantic request part not passed validation." }
        error: { type: string, description: "The validation error occurred." }
    AppApiExceptionResponseV1:
      type: object
      additionalProperties: false
      description: Representation of exception occurred in REST API format.
      required: [ code, message, timestamp, correlationId ]
      properties:
        code: { type: string, description: "The code of exception occurred." }
        message: { type: string, description: "The message of exception occurred." }
        timestamp: { type: string, format: date-time, description: "The timestamp exception occurred." }
        correlationId: { type: string, description: "The log stream UUID." }